requirements-dev.txt
# Testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
pytest-xdist>=3.5
respx>=0.21
pytest-playwright>=0.5

# Linting & Type Checking
ruff>=0.4
mypy>=1.10
